+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2:
 
 * If an invalid token is returned from token service then manually expire the token.
+* Add graceful handling of token service invocations that return non-successful responses.
 
 ## 0.1: