+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.3
 
 * Decouple from `org.keycloak` libraries and use a simplified mechanism for representing tokens.
 * Allow the specification of `connectTimeout` and `readTimeout` used when attempting to connect to keycloak server.